Refactoring transfer classes
