+++ cyoa/app/models/api_data_transfer/point_loader.rb
@@ -38,3 +38,4 @@ module PointLoader
     Point
   end
 end
+
