+++ cyoa/app/models/forecast_type.rb
@@ -1,4 +1,4 @@
 class ForecastType < ActiveRecord::Base
-  scope :daily, -> { where(forecast_type: '24-hour') }
-  scope :three_hour, -> { where(forecast_type: "3-hour") }
+  # scope :daily, -> { where(forecast_type: '24-hour') }
+  # scope :three_hour, -> { where(forecast_type: "3-hour") }
 end
