+++ cyoa/db/schema.rb
@@ -17,9 +17,11 @@ ActiveRecord::Schema.define(version: 20150209151034) do
   enable_extension "plpgsql"
 
   create_table "forecast_types", force: :cascade do |t|
-    t.string   "forecast_type", null: false
-    t.datetime "created_at",    null: false
-    t.datetime "updated_at",    null: false
+    t.string   "forecast_type",           null: false
+    t.datetime "created_at",              null: false
+    t.datetime "updated_at",              null: false
+    t.datetime "last_refresh_start_time"
+    t.datetime "last_refresh_end_time"
     t.index ["forecast_type"], :name => "unique_forecast_type", :unique => true
   end
 
@@ -33,18 +35,16 @@ ActiveRecord::Schema.define(version: 20150209151034) do
   end
 
   create_table "forecasts", force: :cascade do |t|
-    t.integer  "point_id",                            null: false
-    t.datetime "start_time",                          null: false
-    t.datetime "end_time",                            null: false
+    t.integer  "point_id",                     null: false
+    t.datetime "start_time",                   null: false
+    t.datetime "end_time",                     null: false
     t.integer  "maxt"
     t.integer  "mint"
     t.integer  "cloud_cover"
-    t.string   "icon_link",               limit: 255
-    t.datetime "created_at",                          null: false
-    t.datetime "updated_at",                          null: false
+    t.string   "icon_link",        limit: 255
+    t.datetime "created_at",                   null: false
+    t.datetime "updated_at",                   null: false
     t.integer  "forecast_type_id"
-    t.datetime "last_refresh_begin_time"
-    t.datetime "last_refresh_end_time"
     t.index ["forecast_type_id"], :name => "fk__forecasts_forecast_type_id"
     t.index ["point_id", "start_time", "end_time"], :name => "index_forecasts_on_point_id_and_start_time_and_end_time", :unique => true
     t.index ["point_id"], :name => "fk__forecasts_point_id"
